Refactor models into separate files within module
