+++ models/__init__.py
@@ -1,0 +1,7 @@
+from models.associations import *
+from models.campaign import Campaign
+from models.comment import Comment
+from models.epoch import Epoch
+from models.event import Event
+from models.message import Message
+from models.user import User
